processor/requirements.txt
confluent-kafka==2.1.1
pymongo==4.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0